requirements.txt
Django==2.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
mysqlclient==1.3.12
PyJWT==1.6.1
python-decouple==3.1
pytz==2018.3
